@@ -1,34 +1,18 @@
 import {ModuleWithProviders, NgModule, Optional, SkipSelf} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {NbAuthModule, NbDummyAuthProvider} from '@nebular/auth';
 
 import {throwIfAlreadyLoaded} from './module-import-guard';
-import {DataModule} from './data/data.module';
+import {ServicesModule} from './services/services.module';
 
-const NB_CORE_PROVIDERS = [
-  ...DataModule.forRoot().providers,
-  ...NbAuthModule.forRoot({
-    providers: {
-      email: {
-        service: NbDummyAuthProvider,
-        config: {
-          delay: 3000,
-          login: {
-            rememberMe: true,
-          },
-        },
-      },
-    },
-  }).providers,
+export const NB_CORE_PROVIDERS = [
+  ...ServicesModule.forRoot().providers,
 ];
 
 @NgModule({
   imports: [
     CommonModule,
   ],
-  exports: [
-    NbAuthModule,
-  ],
+  exports: [],
   declarations: [],
 })
 export class CoreModule {
