@@ -1,10 +1,7 @@
-import {ModuleWithProviders, NgModule} from '@angular/core';
-import {CommonModule} from '@angular/common';
-
-import {StateService} from './state.service';
+import { NgModule, ModuleWithProviders } from '@angular/core';
+import { CommonModule } from '@angular/common';
 
 const SERVICES = [
-  StateService,
 ];
 
 @NgModule({
