@@ -1,2 +1,1 @@
-export * from './sample/sample.layout';
-export * from './sample/donation-modal/donation-modal.component';
+export * from './rmca/rmca.layout';
