@@ -1,31 +1,19 @@
 import {ExtraOptions, RouterModule, Routes} from '@angular/router';
 import {NgModule} from '@angular/core';
 
-import {NeedLoginGuard} from './auth/guards/needLogin.guard';
-
 const routes: Routes = [
   {
     path: 'pages',
     loadChildren: 'app/pages/pages.module#PagesModule',
-    canActivate: [NeedLoginGuard],
-  },
-  {
+  }, {
     path: 'auth',
     loadChildren: 'app/auth/auth.module#AuthModule',
-  },
-  {
+  }, {
     path: 'callback',
     loadChildren: 'app/callback/callback.module#CallbackModule',
   },
-  {
-    path: '',
-    redirectTo: 'pages',
-    pathMatch: 'full',
-  },
-  {
-    path: '**',
-    redirectTo: 'pages',
-  },
+  {path: '', redirectTo: 'pages', pathMatch: 'full'},
+  {path: '**', redirectTo: 'pages'},
 ];
 
 const config: ExtraOptions = {
