@@ -1,13 +1,13 @@
 import {Component, OnInit} from '@angular/core';
 import {NgbModal} from '@ng-bootstrap/ng-bootstrap';
 
-import {NoticeService} from '../../../../@system/notice/notice.service';
+import {NoticeService} from '../../../../@core/services/notice.service';
 import {DashboardService} from '../../../dashboard/dashboard.service';
 
 import {FinanceAddModalComponent} from './finance-add-modal/finance-add-modal.component';
 import {FinanceDetailModalComponent} from './finance-detail-modal/finance-detail-modal.component';
 import {Finance} from '../../../../@model/admin/server/finacne/finance.interface';
-import {FinanceTypeEnum} from "../../../../@model/admin/server/finacne/finance-type.enum";
+import {FinanceTypeEnum} from '../../../../@model/admin/server/finacne/finance-type.enum';
 
 @Component({
   styleUrls: ['./finance.component.scss'],
