@@ -1,6 +1,6 @@
 import {Component, OnInit} from '@angular/core';
 
-import {NoticeService} from '../../../../@system/notice/notice.service';
+import {NoticeService} from '../../../../@core/services/notice.service';
 import {ServerService} from '../server.service';
 
 @Component({
