@@ -1,10 +1,10 @@
 import {AfterViewInit, Component, Input} from '@angular/core';
-import * as skinview3d from '../../../../../assets/vendors/skinview3d.babel.js';
+import {RotatingAnimation, SkinViewer} from 'skinview3d';
 
 import {Role} from '../../../../@model/player/role/role.interface';
 
 @Component({
-  selector: 'rmca-skin-viewer',
+  selector: 'ngx-skin-viewer',
   styleUrls: ['./skin-viewer.component.scss'],
   templateUrl: './skin-viewer.component.html',
 })
@@ -18,15 +18,11 @@ export class SkinViewerComponent implements AfterViewInit {
   }
 
   public ngAfterViewInit(): void {
-    let skinViewer = new skinview3d.SkinViewer({
+    const skinViewer = new SkinViewer({
       domElement: document.getElementById(this.random),
-      slim: this.role.userModel == 'alex',
       skinUrl: this.role.skin,
       capeUrl: this.role.cape,
-      animation: skinview3d.WalkAnimation
+      animation: RotatingAnimation,
     });
-
-    let control = new skinview3d.SkinControl(skinViewer);
-    control.enableAnimationControl = true;
   }
 }
