@@ -18,9 +18,11 @@
  *
  * Learn more in https://angular.io/docs/ts/latest/guide/browser-support.html
  */
+
 /***************************************************************************************************
  * BROWSER POLYFILLS
  */
+
 /** IE9, IE10 and IE11 requires all of the following polyfills. **/
 import 'core-js/es6/symbol';
 import 'core-js/es6/object';
@@ -42,22 +44,25 @@ import 'web-animations-js'; // Run `npm install --save web-animations-js`.
 /** Evergreen browsers require these. **/
 import 'core-js/es6/reflect';
 import 'core-js/es7/reflect';
-/** ALL Firefox browsers require the following to support `@angular/animation`. **/
-// import 'web-animations-js';  // Run `npm install --save web-animations-js`.
 /***************************************************************************************************
  * Zone JS is required by Angular itself.
  */
 import 'zone.js/dist/zone'; // Included with Angular CLI.
-/***************************************************************************************************
- * APPLICATION IMPORTS
- */
 /**
  * Date, currency, decimal and percent pipes.
  * Needed for: All but Chrome, Firefox, Edge, IE11 and Safari 10
  */
-import 'intl'; // Run `npm install --save intl`.
-import 'intl/locale-data/jsonp/en';
 import 'core-js/es7/array';
+import 'core-js/es7/object';
+
+
+/** ALL Firefox browsers require the following to support `@angular/animation`. **/
+// import 'web-animations-js';  // Run `npm install --save web-animations-js`.
+
+
+/***************************************************************************************************
+ * APPLICATION IMPORTS
+ */
 
 if (typeof SVGElement.prototype.contains === 'undefined') {
   SVGElement.prototype.contains = HTMLDivElement.prototype.contains;
